+++ common/src/main/java/io/github/mortuusars/exposure/gui/screen/PhotographScreen.java
@@ -8,11 +8,13 @@ import com.mojang.datafixers.util.Either;
 import io.github.mortuusars.exposure.Config;
 import io.github.mortuusars.exposure.Exposure;
 import io.github.mortuusars.exposure.ExposureClient;
-import io.github.mortuusars.exposure.camera.capture.component.FileSaveComponent;
 import io.github.mortuusars.exposure.camera.infrastructure.FrameData;
+import io.github.mortuusars.exposure.data.storage.ExposureExporter;
 import io.github.mortuusars.exposure.gui.screen.element.Pager;
 import io.github.mortuusars.exposure.item.PhotographItem;
+import io.github.mortuusars.exposure.render.PhotographRenderProperties;
 import io.github.mortuusars.exposure.render.PhotographRenderer;
+import io.github.mortuusars.exposure.util.ClientsideWorldNameGetter;
 import io.github.mortuusars.exposure.util.ItemAndStack;
 import io.github.mortuusars.exposure.util.PagingDirection;
 import net.minecraft.client.Minecraft;
@@ -183,13 +185,21 @@ public class PhotographScreen extends ZoomableScreen {
             return;
 
         idOrTexture.ifLeft(id -> {
-            if (savedExposures.contains(id))
+            PhotographRenderProperties properties = PhotographRenderProperties.get(photograph.getStack());
+            String filename = properties != PhotographRenderProperties.DEFAULT ? id + "_" + properties.getId() : id;
+
+            if (savedExposures.contains(filename))
                 return;
 
             ExposureClient.getExposureStorage().getOrQuery(id).ifPresent(exposure -> {
-                savedExposures.add(id);
-                new Thread(() -> FileSaveComponent.withDefaultFolders(id)
-                        .save(exposure.getPixels(), exposure.getWidth(), exposure.getHeight(), exposure.getProperties()), "ExposureSaving").start();
+                savedExposures.add(filename);
+
+                new Thread(() -> new ExposureExporter(filename)
+                        .withDefaultFolder()
+                        .organizeByWorld(Config.Client.EXPOSURE_SAVING_LEVEL_SUBFOLDER.get(), ClientsideWorldNameGetter::getWorldName)
+                        .withModifier(properties.getModifier())
+                        .withSize(Config.Client.EXPOSURE_SAVING_SIZE.get())
+                        .save(exposure), "ExposureSaving").start();
             });
         });
     }