+++ forge/src/main/java/io/github/mortuusars/exposure/forge/event/CommonEvents.java
@@ -4,6 +4,8 @@ import io.github.mortuusars.exposure.Exposure;
 import io.github.mortuusars.exposure.command.ExposureCommand;
 import io.github.mortuusars.exposure.command.ShaderCommand;
 import io.github.mortuusars.exposure.command.TestCommand;
+import io.github.mortuusars.exposure.command.argument.ExposureLookArgument;
+import io.github.mortuusars.exposure.command.argument.ExposureSizeArgument;
 import io.github.mortuusars.exposure.command.argument.ShaderLocationArgument;
 import io.github.mortuusars.exposure.data.Lenses;
 import io.github.mortuusars.exposure.data.LensesDataLoader;
@@ -69,6 +71,8 @@ public class CommonEvents {
         @SubscribeEvent
         public static void registerCommands(RegisterCommandsEvent event) {
             ArgumentTypeInfos.registerByClass(ShaderLocationArgument.class, SingletonArgumentInfo.contextFree(ShaderLocationArgument::new));
+            ArgumentTypeInfos.registerByClass(ExposureSizeArgument.class, SingletonArgumentInfo.contextFree(ExposureSizeArgument::new));
+            ArgumentTypeInfos.registerByClass(ExposureLookArgument.class, SingletonArgumentInfo.contextFree(ExposureLookArgument::new));
 
             ExposureCommand.register(event.getDispatcher());
             ShaderCommand.register(event.getDispatcher());